Use base extractor.
