+++ stko/calculators/extractors/__init__.py
@@ -1,2 +1,3 @@
+from .extractor import *  # noqa
 from .xtb_extractor import *  # noqa
 from .orca_extractor import *  # noqa
