+++ stko/calculators/extractors/orca_extractor.py
@@ -9,9 +9,10 @@ Class to extract properties from Orca output.
 """
 
 import re
+from .extractor import Extractor
 
 
-class OrcaExtractor:
+class OrcaExtractor(Extractor):
     """
     Extracts properties from Orca 4.2 output files.
 
@@ -32,28 +33,9 @@ class OrcaExtractor:
 
     """
 
-    def __init__(self, output_file):
-        """
-        Initializes :class:`OrcaExtractor`
-
-        Parameters
-        ----------
-        output_file : :class:`str`
-            Output file to extract properties from.
-
-        """
-
-        self.output_file = output_file
-        # Explictly set encoding to UTF-8 because default encoding on
-        # Windows will fail to read the file otherwise.
-        with open(self.output_file, 'r', encoding='UTF-8') as f:
-            self.output_lines = f.readlines()
-
-        self._extract_values()
-
     def _extract_values(self):
         """
-        Extract all properties from xTB output file.
+        Extract all properties from Orca output file.
 
         Returns
         -------
@@ -65,34 +47,9 @@ class OrcaExtractor:
             if self._check_line(line, 'total_energy'):
                 self._extract_total_energy(line)
 
-    def _check_line(self, line, option):
-        """
-        Checks a line for a string based on option.
+    def _properties_dict(self):
 
-        All formatting based on the 190418 version of xTB.
-
-        Parameters
-        ----------
-        line : :class:`str`
-            Line of output file to check.
-
-        option : :class:`str`
-            Define which property and string being checked for.
-            Can be one of ``'total_energy'``.
-
-        Returns
-        -------
-        :class:`bool`
-            Returns ``True`` if the desired string is present.
-
-        """
-
-        options = {
-            'total_energy': 'FINAL SINGLE POINT ENERGY',
-        }
-
-        if options[option] in line:
-            return True
+        return {'total_energy': 'FINAL SINGLE POINT ENERGY'}
 
     def _extract_total_energy(self, line):
         """
